Refactor player-turn-crisis transition to improve flexibility and maintainability
There are a lot of failing tests after this, because one of the additional features was adding Gaius Baltar's Delusional Intuition. In order to avoid having to handle that all over the place, I changed the fixture to use Tom Zarek instead of Gaius. Many tests still need to be updated for this change.