Fix a bug in the acted logic that was causing issues for sharing bonuses
This was leading to differences in which cards were drawn before and after a player took an action. The exact mechanism isn't clear to me, but somehow dogmaInfo.acting was no longer equal to player, even though they had the same player name. Testing the name explictly, which is a unique id for each player, allows it to work correctly.